requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
apscheduler==3.10.4
pydantic==2.9.2
pydantic-settings==2.5.2
requests==2.32.3
python-dotenv==1.0.1
openai==1.51.2
psycopg[binary]==3.2.10

